+++ pipit-firmware/Lookup.h
@@ -9,6 +9,7 @@
 #include "Key.h"
 
 // Max # of bytes of data in a word/macro lookup array entry.
+//  TODO enforce in config
 #define MAX_LOOKUP_DATA_LENGTH 256
 
 class Lookup{
@@ -23,9 +24,9 @@ private:
   uint8_t lookupChord(const Chord* chord, const LookupsOfSeqType* table, Key* keys_out);
 
   uint8_t readSequence(const uint8_t* seq_lookup,
-                       uint32_t seq_length_in_bits, uint32_t seq_num, Key* keys_out);
+                       uint16_t seq_length_in_bits, uint16_t seq_num, Key* keys_out);
 
-  void getBitArray(bool* bits_out, uint32_t len_bits_out, const uint8_t* start, uint32_t bit_offset);
+  void getBitArray(bool* bits_out, uint16_t len_bits_out, const uint8_t* start, uint32_t bit_offset);
   bool bitToBool(const uint8_t* address, uint32_t bit_offset);
